refactor(query-builder): removed unused query component
