--- frontend/src/container/QueryBuilder/components/Query/Query.styles.scss
@@ -1,8 +1,0 @@
-.qb-search-container {
-	display: block;
-	position: relative;
-}
-
-.qb-container {
-	padding: 0 24px;
-}
