--- frontend/src/container/QueryBuilder/components/Query/index.ts
@@ -1,1 +1,0 @@
-export { Query } from './Query';
