+++ frontend/src/container/QueryBuilder/components/index.ts
@@ -5,4 +5,3 @@ export { Formula } from './Formula';
 export { HavingFilterTag } from './HavingFilterTag';
 export { ListItemWrapper } from './ListItemWrapper';
 export { ListMarker } from './ListMarker';
-export { Query } from './Query';
